pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixigpt"
version = "0.1.0"
description = "Production-grade Python client for the PixiGPT API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "PixiGPT", email = "contact@pixigpt.com"}
]
keywords = ["pixigpt", "llm", "ai", "api", "client"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "urllib3>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/PixiGPT/pixigpt-py"
Documentation = "https://pixigpt.com/docs"
Repository = "https://github.com/PixiGPT/pixigpt-py"
